+++ lib/lennarb.rb
@@ -5,42 +5,93 @@
 
 ENV['RACK_ENV'] ||= 'development'
 
-# Extension for Array class
-#
-require 'lennarb/array_extensions'
-
-# Base class for Lennarb
-#
-require 'lenna/application'
-require 'lennarb/version'
-
 # Core extensions
 #
 require 'pathname'
+require 'rack'
 
-# Zeitwerk
+# Base class for Lennarb
 #
-require 'zeitwerk'
+require_relative 'lennarb/request'
+require_relative 'lennarb/response'
+require_relative 'lennarb/route_node'
+require_relative 'lennarb/version'
 
-# Zeitwerk loader
-#
-Zeitwerk::Loader.new.tap do |loader|
-	loader.inflector.inflect('Version' => 'VERSION')
-	loader.push_dir(__dir__)
-	loader.setup
-end
-
-# Lennarb module
-#
-module Lennarb
-	module_function
-
-	# Lennarb root path
+class Lennarb
+	# Error class
 	#
-	# @return [Pathname] the root path
+	class LennarbError < StandardError; end
+
+	# @attribute [r] root
+	# @return [RouteNode]
 	#
-	def root
-		File.expand_path('..', __dir__)
-		Pathname.new(File.expand_path('..', __dir__))
+	attr_reader :root
+
+	# Initialize the application
+	#
+	# @yield { ... } The application
+	#
+	# @return [Lennarb]
+	#
+	def initialize
+		@root = RouteNode.new
+		yield self if block_given?
+	end
+
+	# Split a path into parts
+	#
+	# @parameter [String] path
+	#
+	# @return [Array] parts. Ex. ['users', ':id']
+	#
+	SplitPath = ->(path) { path.split('/').reject(&:empty?) }
+	private_constant :SplitPath
+
+	# Call the application
+	#
+	# @parameter [Hash] env
+	#
+	# @return [Array] response
+	#
+	def call(env)
+		res         = Response.new
+		http_method = env.fetch('REQUEST_METHOD').to_sym
+		parts       = SplitPath[env.fetch('PATH_INFO')]
+
+		block, params = @root.match_route(parts, http_method)
+		return [404, { 'content-type' => 'text/plain' }, ['Not Found']] unless block
+
+		req = Request.new(env, params)
+		instance_exec(req, res, &block)
+
+		res.finish
+	end
+
+	# Add a routes
+	#
+	# @parameter [String] path
+	# @parameter [Proc] block
+	#
+	# @return [void]
+	#
+	def get(path, &block)    = add_route(path, :GET, block)
+	def post(path, &block)   = add_route(path, :POST, block)
+	def put(path, &block)    = add_route(path, :PUT, block)
+	def patch(path, &block)  = add_route(path, :PATCH, block)
+	def delete(path, &block) = add_route(path, :DELETE, block)
+
+	private
+
+	# Add a route
+	#
+	# @parameter [String] path
+	# @parameter [String] http_method
+	# @parameter [Proc] block
+	#
+	# @return [void]
+	#
+	def add_route(path, http_method, block)
+		parts = SplitPath[path]
+		@root.add_route(parts, http_method, block)
 	end
 end