+++ test/test_helper.rb
@@ -5,10 +5,6 @@
 
 ENV['RACK_ENV'] = 'test'
 
-# Set the console level to fatal
-#
-ENV['CONSOLE_LEVEL'] = 'fatal'
-
 $LOAD_PATH.unshift File.expand_path('../lib', __dir__)
 require 'lennarb'
 
@@ -19,23 +15,3 @@ require 'minitest/autorun'
 # Require Rack::test
 #
 require 'rack/test'
-
-# Default settigns for base class
-#
-module Minitest
-	class Test
-		def teardown
-			Lenna::Middleware::App.instance.reset!
-		end
-	end
-end
-
-# Define the base test for request tests
-#
-class ApplicationRequest < Minitest::Test
-	include Rack::Test::Methods
-
-	def teardown
-		Lenna::Middleware::App.instance.reset!
-	end
-end